Succesfully cut out nodes which have lower conf than the minConf. Still, the linked list of nodes with the same item ID (mapItemNodes) are not properly managed yet, resulted in the same frequentsets as the uncut version.

+++ src/ca/pfv/spmf/algorithms/frequentpatterns/fpgrowth/FPNode.java
@@ -45,7 +45,7 @@ public class FPNode {
 	/**
 	 * Added to store max confidence value
 	 */
-	public int maxConf = 0;
+	public double maxConf = 0;
 	
 	/**
 	 * constructor
